rename ChartView>AreaChartView, use title/data props
